+++ src/client/encryption_socket.rs
@@ -18,7 +18,7 @@ use crate::socket::ToSocketId;
 use crate::std_compat::io::Error;
 use crate::std_compat::io::ErrorKind;
 use crate::std_compat::io::Result;
-use crate::std_compat::net::SocketAddr;
+use core::net::SocketAddr;
 use crate::types::DataBuffer;
 
 pub struct EncryptionBoundSocket<S, Rng> {