+++ src/client/receiver.rs
@@ -32,12 +32,12 @@ use crate::packet::PacketType;
 use crate::socket::Socket;
 use crate::std_compat::io::ErrorKind;
 use crate::std_compat::io::Write;
-use crate::std_compat::net::SocketAddr;
 use crate::string::format_str;
 use crate::time::InstantCallback;
 use crate::types::FilePath;
 use crate::writers::block_writer::BlockWriter;
 use crate::writers::single_block_writer::SingleBlockWriter;
+use core::net::SocketAddr;
 
 cfg_encryption! {
     use crate::client::encryption_socket_builder::create_initial_socket;