+++ std-binary/src/receiver.rs
@@ -30,8 +30,13 @@ where
     W: Write,
     CreateWriter: FnOnce(&FilePath) -> BoxedResult<W>,
 {
-    let socket =
-        create_socket(&config.listen, 1, false, 1).map_err(|e| BinError::from(e.to_string()))?;
+    let listen = if let Some(l) = &config.listen {
+        obtain_listen_socket(l).map_err(|e| BinError::from(e.to_string()))?
+    } else {
+        obtain_listen_socket_based_on_endpoint(&config.endpoint)
+            .map_err(|e| BinError::from(e.to_string()))?
+    };
+    let socket = create_socket(listen, 1, false, 1).map_err(|e| BinError::from(e.to_string()))?;
     // init_logger(socket.local_addr().expect("local address"));
 
     let options = ConnectionOptions {
@@ -63,7 +68,7 @@ where
             .parse()
             .expect("Invalid remote file name"),
     };
-    let client_config = config.try_into(false)?;
+    let client_config = config.try_into(listen, false)?;
     let result = match client_config.encryption_key {
         #[cfg(feature = "encryption")]
         Some(key) => receive_file(