+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All visible changes will be documented here. This project adheres to Semantic Versioning.
 
+## Unreleased
+
+### Fixed
+
+- client can use the same port for new connections immediately
+- incorrect ip format for ipv6 when a new receiver is created
+
+### Changed
+
+- default listen address for receiver and sender: it is based on the endpoint now
+
+### Changed
+
+- windows: always change server port after initial packet
+
 ## [0.7.1] - 2025-08-24
 
 ### Changed